WIP
Proposed a way to scramble the IOIs while keeping segment duration
fixed.

We take the 4 patterns within segment and look at the IOIs. For IOIs
with value 2,3,4 gridpoints, we shorten half and lenghten half with a
noninteger ratio (e.g. Grahn 2009 used 0.3).

We keep IOIs that are 1 gridpoint long intact. This is because are tones
have duration that spans the whole gridIOI. So if we shortened the
inter-onset inteval for these, successive tones would overlap...

I significantly changed the way the sound is synthesized. This should
work for both, standard simple and complex patterns, and also for
nonmetric patterns.

The only thing that must be fixed now is how target-tone F0 is asigned.
I could not understand how this works, so I'll leave it to Ceren.

Everything I did is commented heavily, so should be okay to understand.